backend/requirements.txt
fastapi[all]
httpx
sqlalchemy
pydantic
pyyaml
python-dotenv
